frontend_owner/requirements.txt
streamlit
requests
streamlit-firebase-auth
python-dotenv
pydantic-settings